Add support for managing users.
Also refractors user-specific code to UsersController.
