+++ app/views/layouts/dashboard.html.erb
@@ -22,6 +22,9 @@
 <ul class="col-md-2 col-sm-12 col-xs-12 dashboard-buttons">
   <li id="new-incident-btn"><%= link_to new_incident_path do %><i class="fa fa-exclamation-triangle fa-5x"></i><br />New Incident<% end %></li>
   <li id="history-btn"><%= link_to history_path do %><i class="fa fa-history fa-5x"></i><br />Incident History<% end %></li>
+  <% if current_user.admin? %>
+    <li id="users-btn"><%= link_to manage_users_path do %><i class="fa fa-users fa-5x"></i><br />Manage Users<% end %></li>
+  <% end %>
 </ul>
 
 <div class="col-md-10 col-sm-12 col-xs-12">
