+++ app/views/users/manage.html.erb
@@ -1,0 +1,16 @@
+<h3 class="text-center">Manage users</h3>
+<table class="table table-hover">
+  <thead>
+    <tr><th>#</th><th>Email</th><th>Delete</th>
+  </thead>
+  <tbody>
+    <% User.all.each do |user| %>
+    <tr><td><%= user.id %></td><td><%= user.email %></td><td><%= link_to 'Delete user', "/users/#{user.id}", method: 'delete', class: 'btn btn-sm btn-danger' %></td>
+    <% end %>
+  </tbody>
+</table>
+<br />
+<p align="center">Want to add someone? <%= link_to 'Create a user', new_user_path, class: 'btn btn-md btn-success' %></p>
+<script>
+  $('#users-btn').addClass('active');
+</script>
