+++ config/routes.rb
@@ -17,8 +17,10 @@ Rails.application.routes.draw do
   post 'subscribers/new' => 'subscribers#create', as: 'create_subscriber'
   get 'subscribers/activate/:activation_key' => 'subscribers#activate', as: 'activate_subscriber'
   get 'subscribers/unsubscribe/:activation_key' => 'subscribers#unsubscribe', as: 'unsubscribe_subscriber'
-  get 'users/new' => 'pages#new_user', as: 'new_user'
-  post 'users' => 'pages#create_user'
+  get 'users/new' => 'users#new', as: 'new_user'
+  post 'users' => 'users#create'
+  get 'users/manage' => 'users#manage', as: 'manage_users'
+  delete 'users/:id' => 'users#delete'
   if Rails.env == 'development'
     require 'sidekiq/web'
     mount Sidekiq::Web => '/sidekiq'
